separate out unit test state tests + fix csv fixture tests
